requirements.txt
streamlit==1.25.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==0.27.8
torch==2.0.1
numpy==1.24.3